requirements.txt
torch
numpy
opencv-python
transformers
timm
Pillow
requests
peft>=0.8.2 